pyproject.toml
[tool.isort]
profile = "black"
src_paths = ["app"]

[tool.pycln]
all = true

[tool.poetry]
name = "poukidex"
version = "0.1.0"
description = "Backend of Poukidex app"
authors = ["Thomas Darget <thomasdarget@hotmail.fr>", "Hicham Bakri <hicham.bakri76@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.7"
django-ninja = "0.21.0"
orjson = "3.8.6"
celery = "^5.2.7"
psycopg2-binary = "^2.9.3"
Unidecode = "^1.3.6"
Pillow = "9.4.0"
PyJWT = "^2.5.0"
django-redis = "^5.2.0"
google-cloud-storage = "^2.7.0"
django-cors-headers = "^3.14.0"


[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
